Clone page directory when creating a new task
Before this commit, all kernel tasks shared the same,
single page directory & page tables.

Starting with this commit, when a new kernel task is created,
it gets a new page directory and page tables.

This way, a change in the virtual memory map of one kernel task
would not affect that of another one.

also fix memcpy (oops!)