Scheduler: fix race condition in Scheduler::sleep_ms
previously, if a timer interrupt occured while we were in sleep_ms(),
the task & Scheduler data structures
could be left at an invalid state.
(Specifically, the sleeping task would no be appended to the blocked
tasks list).

This would cause the sleeping task to not be scheduled anymore.

This commit ensures this doesn't happen by disabling interrupts at the
beginning of sleep_ms,
and enabling them at the end.

TODO: when scheduling tasks at a high freqeucny,
there is a heap memory leak.
It might be caused by heap fragmentation in the
kmalloc implementation.